Fix assignment1 issue:  where to develop?
Assignment 1 has a predefined test.js that is used for the assignment.
It is not FOUO and is needed for the .zip and jsbin environments,
therefore it should be in the public git repo.  The final code.js is
FOUO and therefore should be in the FOUO git repo
(exam-assignments-answers).  So we develop both in assignment1-dev and save
the test.js as is and copy the code.js into FOUO/assignment1/code.js and
save it in the FOUO git repo.

The alternatives were:
- to develop in the FOUO folder.  But that means test.js is FOUO.
- to develop in the website folder, but that mean code.js could be
accidentally overwritten.

This solution (assignment1-dev) avoids both problems.